requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.2
chromadb==0.4.5
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.99.1
filelock==3.15.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langsmith==0.1.80
marshmallow==3.21.3
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
onnxruntime==1.18.0
openai==1.34.0
openpyxl==3.1.5
orjson==3.10.5
overrides==7.7.0
packaging==23.2
pandas==2.2.1
posthog==3.5.0
protobuf==5.27.1
pulsar-client==3.5.0
pydantic==1.10.17
pydantic_core==2.20.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.27.0
sympy==1.12.1
tenacity==8.4.1
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
